Updated LR plus training and test data
The regression code runs on regression_training_data.csv, which was the
file used in the evaluation to train the model which was then tested on
regression_test_data.csv.

As some of the training data had missing values, two ressors were
dropped in the evaluation. The regression code has been updated to
reflect this and to note that these can be reintroduced in the future.