Check to see if the score has changed before writing
Amended the code so that, for each row, nothing is written to the
database if the new calculated score is not different from the score
already in the database for that row. In testing, that speeds things up
by over 60% if the databases already populated with scores.

(If the model is changed, all the scores will be rewritten because they
will all be different.)